// src/models/ICountry.ts
export interface NativeName {
    msa: Msa;
    ces: Ces;
    slk: Slk;
    fra: Fra;
    cat: Cat;
    por: Por;
    rus: Rus;
    uzb: Uzb;
    eng: Eng;
    zho: Zho;
    tvl: Tvl;
    hrv: Hrv;
    ara: Ara;
    dan: Dan;
    nau: Nau;
    run: Run;
    sin: Sin;
    tam: Tam;
    spa: Spa;
    tgk: Tgk;
    nrf: Nrf;
    deu: Deu;
    nld: Nld;
    gle: Gle;
    swa: Swa;
    mya: Mya;
    tha: Tha;
    kon: Kon;
    lin: Lin;
    nor: Nor;
    aze: Aze;
    kin: Kin;
    sqi: Sqi;
    srp: Srp;
    hun: Hun;
    hmo: Hmo;
    tpi: Tpi;
    nno: Nno;
    nob: Nob;
    smi: Smi;
    pap: Pap;
    fao: Fao;
    mkd: Mkd;
    kal: Kal;
    ber: Ber;
    mey: Mey;
    tet: Tet;
    ron: Ron;
    pov: Pov;
    tir: Tir;
    smo: Smo;
    tkl: Tkl;
    dzo: Dzo;
    lao: Lao;
    afr: Afr;
    nbl: Nbl;
    nso: Nso;
    sot: Sot;
    ssw: Ssw;
    tsn: Tsn;
    tso: Tso;
    ven: Ven;
    xho: Xho;
    zul: Zul;
    kor: Kor;
    prs: Prs;
    pus: Pus;
    tuk: Tuk;
    jam: Jam;
    gil: Gil;
    swe: Swe;
    fij: Fij;
    hif: Hif;
    crs: Crs;
    arc: Arc;
    ckb: Ckb;
    zdj: Zdj;
    ita: Ita;
    grn: Grn;
    glv: Glv;
    pol: Pol;
    isl: Isl;
    mri: Mri;
    nzs: Nzs;
    vie: Vie;
    tur: Tur;
    div: Div;
    her: Her;
    hgm: Hgm;
    kwn: Kwn;
    loz: Loz;
    ndo: Ndo;
    kir: Kir;
    fin: Fin;
    lat: Lat;
    mlt: Mlt;
    pih: Pih;
    pau: Pau;
    bwg: Bwg;
    kck: Kck;
    khi: Khi;
    ndc: Ndc;
    nde: Nde;
    nya: Nya;
    sna: Sna;
    toi: Toi;
    zib: Zib;
    urd: Urd;
    ukr: Ukr;
    ben: Ben;
    lav: Lav;
    niu: Niu;
    aym: Aym;
    que: Que;
    khm: Khm;
    lit: Lit;
    hat: Hat;
    slv: Slv;
    cal: Cal;
    cha: Cha;
    mfe: Mfe;
    bis: Bis;
    ltz: Ltz;
    ton: Ton;
    ell: Ell;
    som: Som;
    sag: Sag;
    bos: Bos;
    kat: Kat;
    jpn: Jpn;
    nep: Nep;
    rar: Rar;
    nfr: Nfr;
    fas: Fas;
    bar: Bar;
    lua: Lua;
    bel: Bel;
    est: Est;
    mlg: Mlg;
    mon: Mon;
    mah: Mah;
    heb: Heb;
    ind: Ind;
    bjz: Bjz;
    hye: Hye;
    fil: Fil;
    bul: Bul;
    hin: Hin;
    amh: Amh;
    gsw: Gsw;
    roh: Roh;
    cnr: Cnr;
    kaz: Kaz;
}

export interface Name {
    common: string;
    official: string;
    nativeName: NativeName;
}

export interface BND {
    name: string;
    symbol: string;
}

export interface SGD {
    name: string;
    symbol: string;
}

export interface CZK {
    name: string;
    symbol: string;
}

export interface XOF {
    name: string;
    symbol: string;
}

export interface EUR {
    name: string;
    symbol: string;
}

export interface XPF {
    name: string;
    symbol: string;
}

export interface BRL {
    name: string;
    symbol: string;
}

export interface UZS {
    name: string;
    symbol: string;
}

export interface SLL {
    name: string;
    symbol: string;
}

export interface HKD {
    name: string;
    symbol: string;
}

export interface AUD {
    name: string;
    symbol: string;
}

export interface TVD {
    name: string;
    symbol: string;
}

export interface QAR {
    name: string;
    symbol: string;
}

export interface DKK {
    name: string;
    symbol: string;
}

export interface BIF {
    name: string;
    symbol: string;
}

export interface LKR {
    name: string;
    symbol: string;
}

export interface YER {
    name: string;
    symbol: string;
}

export interface USD {
    name: string;
    symbol: string;
}

export interface LBP {
    name: string;
    symbol: string;
}

export interface GBP {
    name: string;
    symbol: string;
}

export interface SHP {
    name: string;
    symbol: string;
}

export interface TJS {
    name: string;
    symbol: string;
}

export interface CNY {
    name: string;
    symbol: string;
}

export interface KWD {
    name: string;
    symbol: string;
}

export interface CUC {
    name: string;
    symbol: string;
}

export interface CUP {
    name: string;
    symbol: string;
}

export interface UYU {
    name: string;
    symbol: string;
}

export interface JEP {
    name: string;
    symbol: string;
}

export interface XAF {
    name: string;
    symbol: string;
}

export interface UGX {
    name: string;
    symbol: string;
}

export interface TZS {
    name: string;
    symbol: string;
}

export interface GNF {
    name: string;
    symbol: string;
}

export interface JOD {
    name: string;
    symbol: string;
}

export interface MMK {
    name: string;
    symbol: string;
}

export interface THB {
    name: string;
    symbol: string;
}

export interface ANG {
    name: string;
    symbol: string;
}

export interface AZN {
    name: string;
    symbol: string;
}

export interface RWF {
    name: string;
    symbol: string;
}

export interface HUF {
    name: string;
    symbol: string;
}

export interface PGK {
    name: string;
    symbol: string;
}

export interface CAD {
    name: string;
    symbol: string;
}

export interface NOK {
    name: string;
    symbol: string;
}

export interface DZD {
    name: string;
    symbol: string;
}
// ...
export interface Currencies {
    BND: BND;
    SGD: SGD;
    CZK: CZK;
    XOF: XOF;
    EUR: EUR;
    XPF: XPF;
    BRL: BRL;
    UZS: UZS;
    SLL: SLL;
    HKD: HKD;
    AUD: AUD;
    TVD: TVD;
    QAR: QAR;
    DKK: DKK;
    BIF: BIF;
    LKR: LKR;
    YER: YER;
    USD: USD;
    LBP: LBP;
    GBP: GBP;
    SHP: SHP;
    TJS: TJS;
    CNY: CNY;
    KWD: KWD;
    CUC: CUC;
    CUP: CUP;
    UYU: UYU;
    JEP: JEP;
    XAF: XAF;
    UGX: UGX;
    TZS: TZS;
    GNF: GNF;
    JOD: JOD;
    MMK: MMK;
    THB: THB;
    ANG: ANG;
    AZN: AZN;
    RWF: RWF;
    HUF: HUF;
    PGK: PGK;
    CAD: CAD;
    NOK: NOK;
    DZD: DZD;
    EGP: EGP;
    RSD: RSD;
    FOK: FOK;
    MKD: MKD;
    OMR: OMR;
    MAD: MAD;
    MRU: MRU;
    NZD: NZD;
    RON: RON;
    DOP: DOP;
    ERN: ERN;
    BTN: BTN;
    INR: INR;
    LAK: LAK;
    KES: KES;
    ZAR: ZAR;
    KRW: KRW;
    AFN: AFN;
    JMD: JMD;
    KID: KID;
    BMD: BMD;
    FJD: FJD;
    SCR: SCR;
    IQD: IQD;
    KMF: KMF;
    TWD: TWD;
    PYG: PYG;
    IMP: IMP;
    LRD: LRD;
    PLN: PLN;
    COP: COP;
    NGN: NGN;
    SRD: SRD;
    ISK: ISK;
    MZN: MZN;
    AED: AED;
    LYD: LYD;
    PAB: PAB;
    VND: VND;
    CVE: CVE;
    SZL: SZL;
    TRY: TRY;
    SSP: SSP;
    BHD: BHD;
    MVR: MVR;
    CRC: CRC;
    NAD: NAD;
    KGS: KGS;
    SAR: SAR;
    AOA: AOA;
    ZWL: ZWL;
    PKR: PKR;
    UAH: UAH;
    HNL: HNL;
    KPW: KPW;
    WST: WST;
    GHS: GHS;
    BDT: BDT;
    PEN: PEN;
    SYP: SYP;
    BOB: BOB;
    KYD: KYD;
    KHR: KHR;
    NIO: NIO;
    MOP: MOP;
    XCD: XCD;
    HTG: HTG;
    ILS: ILS;
    BSD: BSD;
    DJF: DJF;
    AWG: AWG;
    MUR: MUR;
    VUV: VUV;
    RUB: RUB;
    TOP: TOP;
    SOS: SOS;
    LSL: LSL;
    GMD: GMD;
    TND: TND;
    ARS: ARS;
    CLP: CLP;
    BAM: BAM;
    GEL: GEL;
    SEK: SEK;
    JPY: JPY;
    NPR: NPR;
    CHF: CHF;
    MWK: MWK;
    GIP: GIP;
    CKD: CKD;
    GGP: GGP;
    IRR: IRR;
    FKP: FKP;
    CDF: CDF;
    MXN: MXN;
    TTD: TTD;
    BYN: BYN;
    MGA: MGA;
    SBD: SBD;
    TMT: TMT;
    SDG: SDG;
    MNT: MNT;
    BWP: BWP;
    MDL: MDL;
    ZMW: ZMW;
    GTQ: GTQ;
    VES: VES;
    IDR: IDR;
    BZD: BZD;
    GYD: GYD;
    AMD: AMD;
    STN: STN;
    PHP: PHP;
    BGN: BGN;
    MYR: MYR;
    BBD: BBD;
    ETB: ETB;
    ALL: ALL;
    KZT: KZT;
}

export interface Idd {
    root: string;
    suffixes: string[];
}

export interface Languages {
    msa: string;
    ces: string;
    slk: string;
    fra: string;
    cat: string;
    por: string;
    rus: string;
    uzb: string;
    eng: string;
    zho: string;
    tvl: string;
    hrv: string;
    ara: string;
    dan: string;
    nau: string;
    run: string;
    sin: string;
    tam: string;
    spa: string;
    tgk: string;
    nrf: string;
    deu: string;
    nld: string;
    gle: string;
    swa: string;
    mya: string;
    tha: string;
    kon: string;
    lin: string;
    nor: string;
    aze: string;
    kin: string;
    sqi: string;
    srp: string;
    hun: string;
    hmo: string;
    tpi: string;
    nno: string;
    nob: string;
    smi: string;
    pap: string;
    fao: string;
    mkd: string;
    kal: string;
    ber: string;
    mey: string;
    tet: string;
    ron: string;
    pov: string;
    tir: string;
    smo: string;
    tkl: string;
    dzo: string;
    lao: string;
    afr: string;
    nbl: string;
    nso: string;
    sot: string;
    ssw: string;
    tsn: string;
    tso: string;
    ven: string;
    xho: string;
    zul: string;
    kor: string;
    prs: string;
    pus: string;
    tuk: string;
    jam: string;
    gil: string;
    swe: string;
    fij: string;
    hif: string;
    crs: string;
    arc: string;
    ckb: string;
    zdj: string;
    ita: string;
    grn: string;
    glv: string;
    pol: string;
    isl: string;
    mri: string;
    nzs: string;
    vie: string;
    tur: string;
    div: string;
    her: string;
    hgm: string;
    kwn: string;
    loz: string;
    ndo: string;
    kir: string;
    fin: string;
    lat: string;
    mlt: string;
    pih: string;
    pau: string;
    bwg: string;
    kck: string;
    khi: string;
    ndc: string;
    nde: string;
    nya: string;
    sna: string;
    toi: string;
    zib: string;
    urd: string;
    ukr: string;
    ben: string;
    lav: string;
    niu: string;
    aym: string;
    que: string;
    khm: string;
    lit: string;
    hat: string;
    slv: string;
    cal: string;
    cha: string;
    mfe: string;
    bis: string;
    ltz: string;
    ton: string;
    ell: string;
    som: string;
    sag: string;
    bos: string;
    kat: string;
    jpn: string;
    nep: string;
    rar: string;
    nfr: string;
    fas: string;
    bar: string;
    lua: string;
    bel: string;
    est: string;
    mlg: string;
    mon: string;
    mah: string;
    heb: string;
    ind: string;
    bjz: string;
    hye: string;
    fil: string;
    bul: string;
    hin: string;
    amh: string;
    gsw: string;
    roh: string;
    cnr: string;
    kaz: string;
}
// ...
export interface Translations {
    ara: Ara2;
    bre: Bre;
    ces: Ces2;
    cym: Cym;
    deu: Deu2;
    est: Est2;
    fin: Fin2;
    fra: Fra2;
    hrv: Hrv2;
    hun: Hun2;
    ita: Ita2;
    jpn: Jpn2;
    kor: Kor2;
    nld: Nld2;
    per: Per;
    pol: Pol2;
    por: Por2;
    rus: Rus2;
    slk: Slk2;
    spa: Spa2;
    srp: Srp2;
    swe: Swe2;
    tur: Tur2;
    urd: Urd2;
    zho: Zho2;
}

export interface Eng2 {
    f: string;
    m: string;
}

export interface Fra3 {
    f: string;
    m: string;
}

export interface Demonyms {
    eng: Eng2;
    fra: Fra3;
}

export interface Maps {
    googleMaps: string;
    openStreetMaps: string;
}

export interface Car {
    signs: string[];
    side: string;
}

export interface Flags {
    png: string;
    svg: string;
    alt: string;
}

export interface CoatOfArms {
    png: string;
    svg: string;
}

export interface CapitalInfo {
    latlng: number[];
}

export interface PostalCode {
    format: string;
    regex: string;
}

export interface Gini {
    2018: number;
    2011?: number;
    2019?: number;
    2003?: number;
    2010?: number;
    2012?: number;
    2013?: number;
    2016?: number;
    2014?: number;
    2015?: number;
    2017?: number;
    2005?: number;
    2009?: number;
    2004?: number;
    2006?: number;
    1999?: number;
    2008?: number;
    1992?: number;
    1998?: number;
}

export interface ICountry {
    name: Name;
    tld: string[];
    cca2: string;
    ccn3: string;
    cca3: string;
    cioc: string;
    independent: boolean;
    status: string;
    unMember: boolean;
    currencies: Currencies;
    idd: Idd;
    capital: string[];
    altSpellings: string[];
    region: string;
    subregion: string;
    languages: Languages;
    translations: Translations;
    latlng: number[];
    landlocked: boolean;
    borders: string[];
    area: number;
    demonyms: Demonyms;
    flag: string;
    maps: Maps;
    population: number;
    fifa: string;
    car: Car;
    timezones: string[];
    continents: string[];
    flags: Flags;
    coatOfArms: CoatOfArms;
    startOfWeek: string;
    capitalInfo: CapitalInfo;
    postalCode: PostalCode;
    gini: Gini;
}